refactor: extract dataclasses into models module
