+++ game/population.py
@@ -5,7 +5,7 @@ import random
 from typing import List, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from .game import Faction
+    from .models import Faction
 
 
 @dataclass
